refactor: move all existing resolver to legacy
